requirements-test.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile requirements-test.in
#
appdirs==1.4.4            # via black
attrs==20.2.0             # via pytest
autopep8==1.5.4           # via -r requirements-test.in
black==20.8b1             # via -r requirements-test.in
click==7.1.2              # via black
coverage==5.3             # via pytest-cov
flake8==3.8.3             # via -r requirements-test.in
iniconfig==1.0.1          # via pytest
isort==5.5.2              # via -r requirements-test.in
mccabe==0.6.1             # via flake8
more-itertools==8.5.0     # via pytest
mypy-extensions==0.4.3    # via black
packaging==20.4           # via pytest
pathspec==0.8.0           # via black
pluggy==0.13.1            # via pytest
py==1.9.0                 # via pytest
pycodestyle==2.6.0        # via autopep8, flake8
pyflakes==2.2.0           # via flake8
pyparsing==2.4.7          # via packaging
pytest-cov==2.10.1        # via -r requirements-test.in
pytest-django==3.9.0      # via -r requirements-test.in
pytest==6.0.2             # via -r requirements-test.in, pytest-cov, pytest-django
regex==2020.7.14          # via black
six==1.15.0               # via packaging
toml==0.10.1              # via autopep8, black, pytest
typed-ast==1.4.1          # via black
typing-extensions==3.7.4.3  # via black
